refactor(community): communityBoardView 폴더 이동
